Any Colour You Like

+++ src/debugWindow.cpp
@@ -25,12 +25,12 @@ void DebugInterface::init(){
     wbkgd(m_mainWindow->get(), COLOR_PAIR(1)); // Set the background color accordingly
     std::string welcome = "Welcome to Nostradamus OS";
     
-    mvwprintw(m_mainWindow->get(), 8, (m_sizeX - welcome.size())/2, "%s", welcome.c_str());
-    mvwprintw(m_mainWindow->get(), 9,(m_sizeX - global_version_string.size())/2,"%s", global_version_string.c_str());
+    mvwprintw(m_mainWindow->get(), 8, (int)(m_sizeX - welcome.size())/2, "%s", welcome.c_str());
+    mvwprintw(m_mainWindow->get(), 9, (int)(m_sizeX - global_version_string.size())/2,"%s", global_version_string.c_str());
     
     wattron(m_mainWindow->get(), A_BLINK);
     std::string pleasecontinue = "Press Tab to Continue";
-    mvwprintw(m_mainWindow->get(), m_sizeY-5,(m_sizeX - pleasecontinue.size()+1)/2,"%s", pleasecontinue.c_str());
+    mvwprintw(m_mainWindow->get(), m_sizeY-5, (int)(m_sizeX - pleasecontinue.size()+1)/2,"%s", pleasecontinue.c_str());
     wattroff(m_mainWindow->get(), A_BLINK);
     
     wrefresh(m_mainWindow->get());
@@ -46,12 +46,12 @@ void DebugInterface::run(){
     wbkgd(m_mainWindow->get(), COLOR_PAIR(1)); // Set the background color accordingly
     std::string welcome = "Welcome to Nostradamus OS";
     
-    mvwprintw(m_mainWindow->get(), 8, (m_sizeX - welcome.size())/2, "%s", welcome.c_str());
-    mvwprintw(m_mainWindow->get(), 9,(m_sizeX - global_version_string.size())/2,"%s", global_version_string.c_str());
+    mvwprintw(m_mainWindow->get(), 8, (int)(m_sizeX - welcome.size())/2, "%s", welcome.c_str());
+    mvwprintw(m_mainWindow->get(), 9, (int)(m_sizeX - global_version_string.size())/2,"%s", global_version_string.c_str());
     
     wattron(m_mainWindow->get(), A_BLINK);
     std::string pleasecontinue = "Press Tab to Continue";
-    mvwprintw(m_mainWindow->get(), m_sizeY-5,(m_sizeX - pleasecontinue.size()+1)/2,"%s", pleasecontinue.c_str());
+    mvwprintw(m_mainWindow->get(), m_sizeY-5, (int)(m_sizeX - pleasecontinue.size()+1)/2,"%s", pleasecontinue.c_str());
     wattroff(m_mainWindow->get(), A_BLINK);
     
     wrefresh(m_mainWindow->get());
